+++ .eslintrc.cjs
@@ -1,0 +1,1 @@
+module.exports = require("@leomotors/config/eslint").config();
