+++ apps/docs/package.json
@@ -8,7 +8,6 @@
     "lint": "eslint src"
   },
   "devDependencies": {
-    "@cocoa-discord/config": "workspace:0.0.0",
     "@vitejs/plugin-vue-jsx": "3.0.1",
     "@vue/eslint-config-prettier": "7.1.0",
     "@vue/eslint-config-typescript": "11.0.3",
