+++ packages/cocoa-discord/tsconfig.json
@@ -1,7 +1,17 @@
 {
-  "extends": "@cocoa-discord/config/tsconfig-package.json",
   "compilerOptions": {
-    "outDir": "./dist"
+    "target": "esnext",
+    "module": "esnext",
+    "declaration": true,
+    "sourceMap": true,
+    "esModuleInterop": true,
+    "forceConsistentCasingInFileNames": true,
+    "strict": true,
+    "skipLibCheck": true,
+    "moduleResolution": "nodenext",
+    "experimentalDecorators": true,
+    "noUncheckedIndexedAccess": true,
+    "outDir": "dist"
   },
   "include": ["./src/**/*"]
 }
