+++ packages/cocoa-discord/typedoc.json
@@ -1,15 +1,15 @@
 {
-    "entryPoints": [
-        "./src/main/index.ts",
-        "./src/message/index.ts",
-        "./src/message/class/index.ts",
-        "./src/meta/index.ts",
-        "./src/slash/index.ts",
-        "./src/slash/class/index.ts",
-        "./src/template/index.ts",
-        "./src/base/index.ts"
-    ],
-    "out": "docs-dist/typedoc",
-    "includeVersion": true,
-    "readme": "README.g.md"
+  "entryPoints": [
+    "./src/main/index.ts",
+    "./src/message/index.ts",
+    "./src/message/class/index.ts",
+    "./src/meta/index.ts",
+    "./src/slash/index.ts",
+    "./src/slash/class/index.ts",
+    "./src/template/index.ts",
+    "./src/base/index.ts"
+  ],
+  "out": "docs-dist/typedoc",
+  "includeVersion": true,
+  "readme": "README.g.md"
 }
