--- packages/config/.eslintrc.js
@@ -1,1 +1,0 @@
-module.exports = require("./eslint-app");
