--- packages/config/.prettierrc.js
@@ -1,1 +1,0 @@
-module.exports = require("./prettier-app");
