--- packages/config/eslint-app.js
@@ -1,30 +1,0 @@
-// @ts-check
-
-/** @type {import("eslint").Linter.Config} */
-const config = {
-    root: true,
-    parser: "@typescript-eslint/parser",
-    plugins: ["@typescript-eslint/eslint-plugin"],
-    extends: [
-        "plugin:@typescript-eslint/recommended",
-        "plugin:prettier/recommended",
-    ],
-    ignorePatterns: ["dist/**/*"],
-    rules: {
-        "prettier/prettier": "warn",
-        eqeqeq: "warn",
-        "@typescript-eslint/no-unused-vars": [
-            "warn",
-            {
-                argsIgnorePattern: "^_",
-                varsIgnorePattern: "^_",
-                caughtErrorsIgnorePattern: "^_",
-            },
-        ],
-        "prefer-const": "warn",
-        "@typescript-eslint/no-non-null-assertion": "off",
-        "@typescript-eslint/no-namespace": "off",
-    },
-};
-
-module.exports = config;
