--- packages/config/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "@cocoa-discord/config",
-  "version": "0.0.0",
-  "private": true,
-  "scripts": {
-    "format": "eslint --fix"
-  },
-  "devDependencies": {
-    "@trivago/prettier-plugin-sort-imports": "4.1.1",
-    "@types/eslint": "8.44.0",
-    "eslint": "8.45.0",
-    "eslint-config-prettier": "8.8.0",
-    "eslint-plugin-prettier": "5.0.0",
-    "prettier": "3.0.0"
-  }
-}
