--- packages/config/tsconfig-package.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "esnext",
-    "module": "commonjs",
-    "declaration": true,
-    "sourceMap": true,
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "strict": true,
-    "skipLibCheck": true,
-    "moduleResolution": "node",
-    "experimentalDecorators": true,
-    "noUncheckedIndexedAccess": true
-  }
-}
