+++ templates/hybrid/tsconfig.json
@@ -1,16 +1,16 @@
 {
-    "compilerOptions": {
-        "target": "esnext",
-        "rootDirs": ["./src"],
-        "outDir": "./dist",
-        "esModuleInterop": true,
-        "forceConsistentCasingInFileNames": true,
-        "strict": true,
-        "skipLibCheck": true,
-        "moduleResolution": "node",
-        "experimentalDecorators": true,
-        "incremental": true,
-        "noUncheckedIndexedAccess": true
-    },
-    "include": ["./src/**/*"]
+  "compilerOptions": {
+    "target": "esnext",
+    "rootDirs": ["./src"],
+    "outDir": "./dist",
+    "esModuleInterop": true,
+    "forceConsistentCasingInFileNames": true,
+    "strict": true,
+    "skipLibCheck": true,
+    "moduleResolution": "node",
+    "experimentalDecorators": true,
+    "incremental": true,
+    "noUncheckedIndexedAccess": true
+  },
+  "include": ["./src/**/*"]
 }
