+++ templates/slash-only/.eslintrc.cjs
@@ -2,28 +2,28 @@
 
 /** @type {import("eslint").Linter.Config} */
 const config = {
-    root: true,
-    parser: "@typescript-eslint/parser",
-    plugins: ["@typescript-eslint/eslint-plugin"],
-    extends: [
-        "plugin:@typescript-eslint/recommended",
-        "plugin:prettier/recommended",
+  root: true,
+  parser: "@typescript-eslint/parser",
+  plugins: ["@typescript-eslint/eslint-plugin"],
+  extends: [
+    "plugin:@typescript-eslint/recommended",
+    "plugin:prettier/recommended",
+  ],
+  rules: {
+    "prettier/prettier": "warn",
+    "@typescript-eslint/no-unused-vars": [
+      "warn",
+      {
+        argsIgnorePattern: "^_",
+        varsIgnorePattern: "^_",
+        caughtErrorsIgnorePattern: "^_",
+      },
     ],
-    rules: {
-        "prettier/prettier": "warn",
-        "@typescript-eslint/no-unused-vars": [
-            "warn",
-            {
-                argsIgnorePattern: "^_",
-                varsIgnorePattern: "^_",
-                caughtErrorsIgnorePattern: "^_",
-            },
-        ],
-        "prefer-const": "warn",
-        "@typescript-eslint/no-non-null-assertion": "off",
-        "@typescript-eslint/no-namespace": "off",
-    },
-    ignorePatterns: ["dist/**/*"],
+    "prefer-const": "warn",
+    "@typescript-eslint/no-non-null-assertion": "off",
+    "@typescript-eslint/no-namespace": "off",
+  },
+  ignorePatterns: ["dist/**/*"],
 };
 
 module.exports = config;
