+++ templates/slash-only/package.json
@@ -1,27 +1,27 @@
 {
-    "name": "@cocoa-discord/template-slash-only",
-    "version": "1.0.0",
-    "description": "Template Discord Bot using Cocoa Discord Utils (Slash Commands Only)",
-    "main": "dist/bot.js",
-    "type": "module",
-    "scripts": {
-        "build": "tsc",
-        "clean": "rm -rf dist/*",
-        "start": "node --es-module-specifier-resolution=node dist/bot.js",
-        "format": "prettier --write \"src/**/*.ts\"",
-        "lint": "eslint src && prettier --check \"src/**/*.ts\""
-    },
-    "devDependencies": {
-        "@trivago/prettier-plugin-sort-imports": "4.1.1",
-        "@typescript-eslint/eslint-plugin": "6.0.0",
-        "@typescript-eslint/parser": "6.0.0",
-        "eslint": "8.45.0",
-        "prettier": "3.0.0",
-        "typescript": "5.1.6"
-    },
-    "dependencies": {
-        "cocoa-discord-utils": "^2.3.0",
-        "discord.js": "^14.11.0",
-        "dotenv": "^16.3.1"
-    }
+  "name": "@cocoa-discord/template-slash-only",
+  "version": "1.0.0",
+  "description": "Template Discord Bot using Cocoa Discord Utils (Slash Commands Only)",
+  "main": "dist/bot.js",
+  "type": "module",
+  "scripts": {
+    "build": "tsc",
+    "clean": "rm -rf dist/*",
+    "start": "node --es-module-specifier-resolution=node dist/bot.js",
+    "format": "prettier --write \"src/**/*.ts\"",
+    "lint": "eslint src && prettier --check \"src/**/*.ts\""
+  },
+  "devDependencies": {
+    "@trivago/prettier-plugin-sort-imports": "4.1.1",
+    "@typescript-eslint/eslint-plugin": "6.0.0",
+    "@typescript-eslint/parser": "6.0.0",
+    "eslint": "8.45.0",
+    "prettier": "3.0.0",
+    "typescript": "5.1.6"
+  },
+  "dependencies": {
+    "cocoa-discord-utils": "^2.3.0",
+    "discord.js": "^14.11.0",
+    "dotenv": "^16.3.1"
+  }
 }
